Add plugable settings. Initial commits for both QSettings and I2C FRAM versions of it.

+++ iodriver.h
@@ -84,7 +84,7 @@ public:
 class Output_Stepper : public Output {
 protected:
     void save();
-    double position;
+	int position;
     int duration;
     bool onOff;
